// backend/src/domain/chat/middlewares/generated/models/AnswerResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CCCC
 * CodeCentric Company Chat
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MessageDto } from './MessageDto';
import {
    MessageDtoFromJSON,
    MessageDtoFromJSONTyped,
    MessageDtoToJSON,
} from './MessageDto';

/**
 * 
 * @export
 * @interface AnswerResponseDto
 */
export interface AnswerResponseDto {
    /**
     * The result messages.
     * @type {MessageDto}
     * @memberof AnswerResponseDto
     */
    result: MessageDto;
    /**
     * The debug information.
     * @type {string}
     * @memberof AnswerResponseDto
     */
    debug?: string;
    /**
     * The summary of the conversation.
     * @type {string}
     * @memberof AnswerResponseDto
     */
    summary?: string;
}

/**
 * Check if a given object implements the AnswerResponseDto interface.
 */
export function instanceOfAnswerResponseDto(value: object): value is AnswerResponseDto {
    if (!('result' in value) || value['result'] === undefined) return false;
    return true;
}

export function AnswerResponseDtoFromJSON(json: any): AnswerResponseDto {
    return AnswerResponseDtoFromJSONTyped(json, false);
}

export function AnswerResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AnswerResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'result': MessageDtoFromJSON(json['result']),
        'debug': json['debug'] == null ? undefined : json['debug'],
        'summary': json['summary'] == null ? undefined : json['summary'],
    };
}

export function AnswerResponseDtoToJSON(value?: AnswerResponseDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'result': MessageDtoToJSON(value['result']),
        'debug': value['debug'],
        'summary': value['summary'],
    };
}